requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
Flask-Migrate==4.0.7
Flask-WTF==1.2.1
Flask-Security-Too==5.4.3
bcrypt==4.1.3
